Move FormattedCst/UnformattedCst to separate file
